+++ getting_start.sh
@@ -3,11 +3,11 @@
 #
 # --- Script Version ---
 # Name    : getting_start.sh
-# Version : 52bb411 (1 commit after this ref)
+# Version : ae6df5d (1 commit after this ref)
 # Author  : IceCodeNew
 # Date    : March 2021
 # Download: https://raw.githubusercontent.com/IceCodeNew/go-collection/master/getting_start.sh
-readonly local_script_version='52bb411'
+readonly local_script_version='ae6df5d'
 
 # IMPORTANT!
 # `apt` does not have a stable CLI interface. Use with caution in scripts.
@@ -106,21 +106,22 @@ install_binaries() {
 
   ########
 
-  tmp_dir=$(mktemp -d)
-  pushd "$tmp_dir" || exit 1
-  download_url="$(curl -sSL -H 'Accept: application/vnd.github.v3+json' \
-    'https://api.github.com/repos/BurntSushi/ripgrep/releases/latest' |
-      grep 'browser_download_url' | cut -d'"' -f4 | grep -iE 'amd64.deb$')"
-  [[ x"${geoip_is_cn:0:1}" = x'y' ]] && download_url=$(echo "$download_url" | sed -E 's!(https://github.com/.+/download/)!https://gh.api.99988866.xyz/\1!g')
-  curl -o 'ripgrep_amd64.deb' -- "$download_url" &&
-    sudo dpkg -i 'ripgrep_amd64.deb' && apt-mark hold ripgrep
-  popd || exit 1
-  /bin/rm -rf "$tmp_dir"
-  dirs -c
-  curl_to_dest "https://github.com/IceCodeNew/rust-collection/releases/latest/download/ripgrep" '/usr/bin/rg'
+  if [[ x"$(echo "${install_ripgrep:=yes}" | cut -c1)" = x'y' ]]; then
+    tmp_dir=$(mktemp -d)
+    pushd "$tmp_dir" || exit 1
+    download_url="$(curl -sSL -H 'Accept: application/vnd.github.v3+json' \
+      'https://api.github.com/repos/BurntSushi/ripgrep/releases/latest' |
+        grep 'browser_download_url' | cut -d'"' -f4 | grep -iE 'amd64.deb$')"
+    [[ x"${geoip_is_cn:0:1}" = x'y' ]] && download_url=$(echo "$download_url" | sed -E 's!(https://github.com/.+/download/)!https://gh.api.99988866.xyz/\1!g')
+    curl -o 'ripgrep_amd64.deb' -- "$download_url" &&
+      sudo dpkg -i 'ripgrep_amd64.deb' && apt-mark hold ripgrep
+    popd || exit 1
+    /bin/rm -rf "$tmp_dir"
+    dirs -c
+    curl_to_dest "https://github.com/IceCodeNew/rust-collection/releases/latest/download/ripgrep" '/usr/bin/rg'
+  fi
 
-  # shellcheck disable=SC2154
-  if [[ x"${install_bat:0:1}" = x'y' ]]; then
+  if [[ x"$(echo "${install_bat:=yes}" | cut -c1)" = x'y' ]]; then
     tmp_dir=$(mktemp -d)
     pushd "$tmp_dir" || exit 1
     download_url="$(curl -sSL -H 'Accept: application/vnd.github.v3+json' \
@@ -140,21 +141,22 @@ install_binaries() {
     curl_to_dest "https://github.com/IceCodeNew/rust-collection/releases/latest/download/bat" '/usr/bin/bat'
   fi
 
-  tmp_dir=$(mktemp -d)
-  pushd "$tmp_dir" || exit 1
-  download_url="$(curl -sSL -H 'Accept: application/vnd.github.v3+json' \
-    'https://api.github.com/repos/sharkdp/fd/releases/latest' |
-      grep 'browser_download_url' | cut -d'"' -f4 | grep -iE 'musl.+amd64.deb$')"
-  [[ x"${geoip_is_cn:0:1}" = x'y' ]] && download_url=$(echo "$download_url" | sed -E 's!(https://github.com/.+/download/)!https://gh.api.99988866.xyz/\1!g')
-  curl -o 'fd-musl_amd64.deb' -- "$download_url" &&
-    sudo dpkg -i 'fd-musl_amd64.deb'
-  popd || exit 1
-  /bin/rm -rf "$tmp_dir"
-  dirs -c
-  curl_to_dest "https://github.com/IceCodeNew/rust-collection/releases/latest/download/fd" '/usr/bin/fd'
+  if [[ x"$(echo "${install_fd:=yes}" | cut -c1)" = x'y' ]]; then
+    tmp_dir=$(mktemp -d)
+    pushd "$tmp_dir" || exit 1
+    download_url="$(curl -sSL -H 'Accept: application/vnd.github.v3+json' \
+      'https://api.github.com/repos/sharkdp/fd/releases/latest' |
+        grep 'browser_download_url' | cut -d'"' -f4 | grep -iE 'musl.+amd64.deb$')"
+    [[ x"${geoip_is_cn:0:1}" = x'y' ]] && download_url=$(echo "$download_url" | sed -E 's!(https://github.com/.+/download/)!https://gh.api.99988866.xyz/\1!g')
+    curl -o 'fd-musl_amd64.deb' -- "$download_url" &&
+      sudo dpkg -i 'fd-musl_amd64.deb'
+    popd || exit 1
+    /bin/rm -rf "$tmp_dir"
+    dirs -c
+    curl_to_dest "https://github.com/IceCodeNew/rust-collection/releases/latest/download/fd" '/usr/bin/fd'
+  fi
 
-  # shellcheck disable=SC2154
-  if [[ x"${install_hexyl:0:1}" = x'y' ]]; then
+  if [[ x"$(echo "${install_hexyl:=yes}" | cut -c1)" = x'y' ]]; then
     tmp_dir=$(mktemp -d)
     pushd "$tmp_dir" || exit 1
     download_url="$(curl -sSL -H 'Accept: application/vnd.github.v3+json' \
@@ -169,8 +171,7 @@ install_binaries() {
     curl_to_dest "https://github.com/IceCodeNew/rust-collection/releases/latest/download/hexyl" '/usr/bin/hexyl'
   fi
 
-  # shellcheck disable=SC2154
-  if [[ x"${install_hugo_extended:0:1}" = x'y' ]] && date +%u | grep -qF '7'; then
+  if [[ x"$(echo "${install_hugo_extended:=no}" | cut -c1)" = x'y' ]]; then
     tmp_dir=$(mktemp -d)
     pushd "$tmp_dir" || exit 1
     download_url="$(curl -sSL -H 'Accept: application/vnd.github.v3+json' \
@@ -190,9 +191,17 @@ install_binaries() {
   curl_to_dest 'https://raw.githubusercontent.com/schollz/croc/master/src/install/bash_autocomplete' '/etc/bash_completion.d/croc' &&
     sudo chmod -x '/etc/bash_completion.d/croc'
 
-  curl_to_dest "https://github.com/IceCodeNew/go-collection/releases/latest/download/shfmt" '/usr/local/bin/shfmt'
+  if [[ x"$(echo "${install_shfmt:=yes}" | cut -c1)" = x'y' ]]; then
+    curl_to_dest "https://github.com/IceCodeNew/go-collection/releases/latest/download/shfmt" '/usr/local/bin/shfmt'
+  else
+    sudo rm '/usr/local/bin/shfmt'
+  fi
 
-  curl_to_dest "https://github.com/IceCodeNew/rust-collection/releases/latest/download/sd" '/usr/local/bin/sd'
+  if [[ x"$(echo "${install_sd:=yes}" | cut -c1)" = x'y' ]]; then
+    curl_to_dest "https://github.com/IceCodeNew/rust-collection/releases/latest/download/sd" '/usr/local/bin/sd'
+  else
+    sudo rm '/usr/local/bin/sd'
+  fi
 
   # shellcheck disable=SC2154
   if [[ x"$(echo "${install_github_release:=no}" | cut -c1)" = x'y' ]]; then
@@ -245,17 +254,21 @@ install_binaries() {
     sudo rm '/usr/local/bin/go-shadowsocks2' '/usr/local/bin/v2ray-plugin'
   fi
 
-  tmp_dir=$(mktemp -d)
-  pushd "$tmp_dir" || exit 1
-  download_url="$(curl -sSL -H 'Accept: application/vnd.github.v3+json' \
-    'https://api.github.com/repos/klzgrad/naiveproxy/releases/latest' |
-      grep 'browser_download_url' | cut -d'"' -f4 | grep -iE 'linux-x64.tar.xz$')"
-  [[ x"${geoip_is_cn:0:1}" = x'y' ]] && download_url=$(echo "$download_url" | sed -E 's!(https://github.com/.+/download/)!https://gh.api.99988866.xyz/\1!g')
-  curl "$download_url" | bsdtar -xf- --strip-components 1
-  # Need glibc runtime.
-  sudo strip './naive' -o '/usr/local/bin/naive'
-  popd || exit 1
-  /bin/rm -rf "$tmp_dir"
+  if [[ x"$(echo "${install_naiveproxy:=yes}" | cut -c1)" = x'y' ]]; then
+    tmp_dir=$(mktemp -d)
+    pushd "$tmp_dir" || exit 1
+    download_url="$(curl -sSL -H 'Accept: application/vnd.github.v3+json' \
+      'https://api.github.com/repos/klzgrad/naiveproxy/releases/latest' |
+        grep 'browser_download_url' | cut -d'"' -f4 | grep -iE 'linux-x64.tar.xz$')"
+    [[ x"${geoip_is_cn:0:1}" = x'y' ]] && download_url=$(echo "$download_url" | sed -E 's!(https://github.com/.+/download/)!https://gh.api.99988866.xyz/\1!g')
+    curl "$download_url" | bsdtar -xf- --strip-components 1
+    # Need glibc runtime.
+    sudo strip './naive' -o '/usr/local/bin/naive'
+    popd || exit 1
+    /bin/rm -rf "$tmp_dir"
+  else
+    sudo rm '/usr/local/bin/naive'
+  fi
 
   # shellcheck disable=SC2154
   if [[ x"$(echo "${install_overmind:=no}" | cut -c1)" = x'y' ]]; then
@@ -279,7 +292,11 @@ install_binaries() {
     sudo rm '/usr/local/bin/chisel'
   fi
 
-  curl_to_dest "https://github.com/IceCodeNew/go-collection/releases/latest/download/got" '/usr/local/bin/got'
+  if [[ x"$(echo "${install_got:=yes}" | cut -c1)" = x'y' ]]; then
+    curl_to_dest "https://github.com/IceCodeNew/go-collection/releases/latest/download/got" '/usr/local/bin/got'
+  else
+    sudo rm '/usr/local/bin/got'
+  fi
 
   # shellcheck disable=SC2154
   if [[ x"$(echo "${install_dive:=yes}" | cut -c1)" = x'y' ]]; then
@@ -295,7 +312,11 @@ install_binaries() {
     sudo rm '/usr/local/bin/duf'
   fi
 
-  curl_to_dest "https://github.com/IceCodeNew/go-collection/releases/latest/download/dnslookup" '/usr/local/bin/dnslookup'
+  if [[ x"$(echo "${install_dnslookup:=yes}" | cut -c1)" = x'y' ]]; then
+    curl_to_dest "https://github.com/IceCodeNew/go-collection/releases/latest/download/dnslookup" '/usr/local/bin/dnslookup'
+  else
+    sudo rm '/usr/local/bin/dnslookup'
+  fi
 
   dog_latest_tag_name="$(curl -sSL -H 'Accept: application/vnd.github.v3+json' \
     'https://api.github.com/repos/ogham/dog/tags?per_page=100' |
@@ -464,27 +485,26 @@ install_binaries() {
   /bin/rm -rf "$tmp_dir"
   dirs -c
 
-  if ! [[ -f /usr/bin/caddy ]] || date +%u | grep -qF '6'; then
-    tmp_dir=$(mktemp -d)
-    pushd "$tmp_dir" || exit 1
-    download_url="$(curl -sSL -H 'Accept: application/vnd.github.v3+json' \
-      'https://api.github.com/repos/caddyserver/caddy/releases/latest' |
-        grep 'browser_download_url' | grep 'linux_amd64.deb' | cut -d'"' -f4)"
-    [[ x"${geoip_is_cn:0:1}" = x'y' ]] && download_url=$(echo "$download_url" | sed -E 's!(https://github.com/.+/download/)!https://gh.api.99988866.xyz/\1!g')
-    curl -o 'caddy_linux_amd64.deb' -- "$download_url" &&
-      sudo dpkg -i 'caddy_linux_amd64.deb' && sudo rm 'caddy_linux_amd64.deb'
-    popd || exit 1
-    /bin/rm -rf "$tmp_dir"
-    dirs -c
-    # shellcheck disable=SC2154
-    if [[ x"${donot_need_caddy_autorun:0:1}" = x'y' ]]; then
-      sudo systemctl disable --now caddy
-    else
-      sudo sed -i -E 's/^:80/:19600/' /etc/caddy/Caddyfile
-    fi
-    sudo rm '/usr/local/bin/caddy' '/usr/local/bin/xcaddy'
-    curl_to_dest "https://github.com/IceCodeNew/go-collection/releases/latest/download/caddy" '/usr/bin/caddy'
+  tmp_dir=$(mktemp -d)
+  pushd "$tmp_dir" || exit 1
+  download_url="$(curl -sSL -H 'Accept: application/vnd.github.v3+json' \
+    'https://api.github.com/repos/caddyserver/caddy/releases/latest' |
+      grep 'browser_download_url' | grep 'linux_amd64.deb' | cut -d'"' -f4)"
+  [[ x"${geoip_is_cn:0:1}" = x'y' ]] && download_url=$(echo "$download_url" | sed -E 's!(https://github.com/.+/download/)!https://gh.api.99988866.xyz/\1!g')
+  curl -o 'caddy_linux_amd64.deb' -- "$download_url" &&
+    sudo dpkg -i 'caddy_linux_amd64.deb' && sudo rm 'caddy_linux_amd64.deb'
+  popd || exit 1
+  /bin/rm -rf "$tmp_dir"
+  dirs -c
+  if [[ x"$(echo "${donot_need_caddy_autorun:=no}" | cut -c1)" = x'y' ]]; then
+    sudo systemctl disable --now caddy
+  else
+    sudo sed -i -E 's/^:80/:19600/' /etc/caddy/Caddyfile
+  fi
+  sudo rm '/usr/local/bin/caddy' '/usr/local/bin/xcaddy'
+  curl_to_dest "https://github.com/IceCodeNew/go-collection/releases/latest/download/caddy" '/usr/bin/caddy'
 
+  if [[ x"$(echo "${install_minify:=no}" | cut -c1)" = x'y' ]]; then
     sudo apt-get update
     sudo apt-get -y install minify
     tmp_dir=$(mktemp -d)
@@ -497,7 +517,13 @@ install_binaries() {
     /bin/rm -rf "$tmp_dir"
     dirs -c
     [[ -f /usr/share/caddy/index.html ]] && minify -o /usr/share/caddy/index.html /usr/share/caddy/index.html
+  else
+    sudo dpkg -P minify
+    sudo rm -rf '/usr/bin/minify' '/etc/bash_completion.d/minify'
   fi
+  sudo rm -f '/usr/share/caddy/index.html' &&
+    sudo mkdir -p '/usr/share/caddy' &&
+    sudo curl -o '/usr/share/caddy/index.html' -- 'https://raw.githubusercontent.com/IceCodeNew/go-collection/master/usr/share/caddy/index.html'
 
   ################
